boom: make command type matching strict
There is an undocumented behaviour that if `boom  [<type>]` is passed a
prefix of a command type, boom will run the first match in the
`_boom_command_types` list.

Historically this has allowed command type prefixes to be used on the
command line, e.g. `# boom ent create`. As well as being undocumented
this behaviour is inconsistent with ArgumentParser subcommands.

Fix `_match_cmd_type()` and enforce strict subcommand matches to aid the
move to ArgumentParser usbcommands.

Resolves: #190

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>